[Update]なぜか消滅した変更を復活、intをunit16に変更

+++ src/core/app/userapp/create_user_app_service.go
@@ -47,8 +47,8 @@ type CreateSkillRequest struct {
 
 type CreateCareerRequest struct {
 	Detail    string
-	StartYear int
-	EndYear   int
+	StartYear uint16
+	EndYear   uint16
 }
 
 type TagParamRequest struct {
