+++ src/core/infra/models/career_model.go
@@ -3,6 +3,6 @@ package models
 type CareerModel struct {
 	CareerID  string `db:"id"`
 	Detail    string `db:"detail"`
-	StartYear int    `db:"start_year"`
-	EndYear   int    `db:"end_year"`
+	StartYear uint16 `db:"start_year"`
+	EndYear   uint16 `db:"end_year"`
 }
